Create README.md
This is a simple game that will ask you to pick 2 numbers. Then it will generate a random number between the 2 numbers you picked. 
You will have to guess what the number is. If the number you inputed is greater than the randome number it will say: "The number is less than that. "
If the number that you guesses is larger then it will say: " The number is greater than that." It will contine until you guess the corrrect number.